refactor form methods to forms.py
